project_06_marketing_agent_slogan_generator/requirements.txt
langchain
langchain-google-genai
python-dotenv
uv
langgraph
requests
langchain-community